Rename MCP settings to Ask Agent connectors
