+++ packages/web/src/app/(app)/settings/accountAskAgent/accountAskAgentPage.test.tsx
@@ -1,0 +1,33 @@
+import { afterEach, describe, expect, test, vi } from 'vitest';
+import { cleanup, render, screen } from '@testing-library/react';
+
+vi.mock('@/app/api/(client)/client', () => ({
+    getMcpServersWithStatus: vi.fn(),
+}));
+vi.mock('@/ee/features/mcp/actions', () => ({
+    disconnectMcpServer: vi.fn(),
+}));
+
+const { AccountAskAgentEmptyState } = await import('./accountAskAgentPage');
+
+afterEach(() => {
+    cleanup();
+});
+
+describe('AccountAskAgentEmptyState', () => {
+    test('points owners to workspace Ask Agent settings', () => {
+        render(<AccountAskAgentEmptyState canManageConnectors={true} />);
+
+        expect(screen.getByText('No connectors configured yet')).toBeTruthy();
+        expect(screen.getByText('Open Workspace Ask Agent to approve connectors for your workspace.')).toBeTruthy();
+        expect(screen.getByRole('link', { name: /Open Workspace Ask Agent/ }).getAttribute('href')).toBe('/settings/workspaceAskAgent');
+    });
+
+    test('tells members to contact an admin', () => {
+        render(<AccountAskAgentEmptyState canManageConnectors={false} />);
+
+        expect(screen.getByText('No connectors available')).toBeTruthy();
+        expect(screen.getByText(/Contact your workspace admin/)).toBeTruthy();
+        expect(screen.queryByRole('link', { name: /Open Workspace Ask Agent/ })).toBeNull();
+    });
+});
