+++ packages/web/src/app/(app)/settings/workspaceAskAgent/page.test.tsx
@@ -20,8 +20,8 @@ vi.mock('@/lib/entitlements', () => ({
 vi.mock('@/middleware/authenticatedPage', () => ({
     authenticatedPage: vi.fn((page: (auth: typeof mocks.authContext) => Promise<React.ReactElement>) => () => page(mocks.authContext)),
 }));
-vi.mock('./mcpConfigurationPage', () => ({
-    McpConfigurationPage: () => <div>MCP configuration client</div>,
+vi.mock('./workspaceAskAgentPage', () => ({
+    WorkspaceAskAgentPage: () => <div>Workspace Ask Agent client</div>,
 }));
 
 const { default: Page } = await import('./page');
@@ -36,11 +36,11 @@ afterEach(() => {
     cleanup();
 });
 
-describe('MCP configuration settings page', () => {
+describe('Ask Agent settings page', () => {
     test('renders the client configuration page when OAuth is available', async () => {
         render(await Page({}));
 
-        expect(screen.getByText('MCP configuration client')).toBeTruthy();
+        expect(screen.getByText('Workspace Ask Agent client')).toBeTruthy();
     });
 
     test('renders the client configuration page when OAuth is unavailable but servers exist for cleanup', async () => {
@@ -49,7 +49,7 @@ describe('MCP configuration settings page', () => {
 
         render(await Page({}));
 
-        expect(screen.getByText('MCP configuration client')).toBeTruthy();
+        expect(screen.getByText('Workspace Ask Agent client')).toBeTruthy();
         expect(mocks.authContext.prisma.mcpServer.count).toHaveBeenCalledWith({
             where: { orgId: 1 },
         });
@@ -60,7 +60,7 @@ describe('MCP configuration settings page', () => {
 
         render(await Page({}));
 
-        expect(screen.getByText('MCP Configuration Is Unavailable')).toBeTruthy();
-        expect(screen.queryByText('MCP configuration client')).toBeNull();
+        expect(screen.getByText('Ask Agent Connectors Are Unavailable')).toBeTruthy();
+        expect(screen.queryByText('Workspace Ask Agent client')).toBeNull();
     });
 });
